Fix format specifier warnings
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@36507 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/c64/patchrom.c
@@ -466,17 +466,19 @@ int patch_rom_idx(int rev)
     uint16_t a;
 
     if (c64rom_get_kernal_chksum_id(&sum, &curr) < 0) {
-        log_error(LOG_DEFAULT, "ROM not patched: Unknown Kernal image.  ID: %d ($%02X) Sum: %d ($%04X).", curr, curr, sum, sum);
+        log_error(LOG_DEFAULT,
+                "ROM not patched: Unknown Kernal image.  ID: %d ($%02X) Sum: %d ($%04X).",
+                curr, (unsigned int)curr, sum, sum);
         return -1;
     }
-    log_verbose("Trying Kernal ROM patch: id:%d/sum:%d to id:%d.", curr, sum, rev);
+    log_verbose("Trying Kernal ROM patch: id:%d/sum:%u to id:%d.", curr, sum, rev);
 
     if (rev == curr) {
         log_verbose("ROM not patched: Already revision #%d.", curr);
         return 0;
     }
     if (rev < 0) {
-        log_error(LOG_DEFAULT, "ROM not patched: Unknown ID: %d ($%02X).", rev, rev);
+        log_error(LOG_DEFAULT, "ROM not patched: Unknown ID: %d ($%02X).", rev, (unsigned int)rev);
         return -1;
     }
